==> multiome/scenicplus/L5_ET/01_prep_atac_data.ipynb <==
# Extract data for scenicplus

suppressPackageStartupMessages({
    library(Seurat)
    library(repr)
    library(patchwork)
    library(ggplot2)
    library(Signac)
    library(tidyverse)
    library(GenomicRanges)
    library(edgeR)
    library(SingleCellExperiment)
    library(Matrix)
    library(scran)
    library(scater)
    library(ggrepel)
    library(fs)
    library(tidyverse)
    library(randomForest)
    library(reticulate)
    library(pheatmap)
    library(gridExtra)
    library(RColorBrewer)
    library(MAST)
    library(data.table)
    library(ComplexHeatmap)
})
options(future.globals.maxSize = Inf)
options(Seurat.object.assay.version = "v5")
options(ggrepel.max.overlaps = Inf)

root_dir <- "/tscc/projects/ps-epigen/users/biy022/biccn/data/SNAREdata/cell_peak_matrix/"
setwd(root_dir)

L5_ET_atac <- readRDS("L5_ET.rds")

meta_data <- read.table(
    "../cell_gene_matrix/20230313_RNA_metadata.xls", sep = "\t", header = TRUE)

write.table(
    meta_data[colnames(L5_ET_atac), ], "../scenicplus/L5_ET/L5_ET_meta_data.tsv",
    col.names = TRUE, row.names = TRUE, quote = FALSE, sep = "\t")

writeMM(L5_ET_atac, "../scenicplus/L5_ET/L5_ET_atac_matrix.mtx")

write.table(
    colnames(L5_ET_atac), "../scenicplus/L5_ET/L5_ET_cell_barcodes.tsv",
    col.names = FALSE, quote = FALSE, sep = "\t", row.names = FALSE)

write.table(
    rownames(L5_ET_atac), "../scenicplus/L5_ET/L5_ET_atac_regions.tsv",
    col.names = FALSE, quote = FALSE, sep = "\t", row.names = FALSE)
