==> multiome/scenicplus/L5_ET/01.1_prep_rna_data.ipynb <==
# Save RNA count to mtx

suppressPackageStartupMessages({
    library(Seurat)
    library(SeuratDisk)
    library(repr)
    library(patchwork)
    library(ggplot2)
    library(Signac)
    library(tidyverse)
    library(GenomicRanges)
    library(Matrix)
    library(ggrepel)
    library(fs)
    library(pheatmap)
    library(gridExtra)
    library(RColorBrewer)
    library(data.table)
    library(ComplexHeatmap)
})
options(future.globals.maxSize = Inf)
options(Seurat.object.assay.version = "v5")
options(ggrepel.max.overlaps = Inf)

root_dir <- "/tscc/projects/ps-epigen/users/biy022/biccn/data/SNAREdata/cell_gene_matrix/"
setwd(root_dir)

L5_ET_rna <- readRDS("L5_ET.rds")

L5_ET_barcodes <- read.table(
    "../scenicplus/L5_ET/L5_ET_cell_barcodes.tsv", 
    header = FALSE, sep = "\t")

L5_ET_rna_filtered <- L5_ET_rna[, L5_ET_barcodes$V1]

writeMM(L5_ET_rna_filtered, "../scenicplus/L5_ET/L5_ET_rna_matrix.mtx")

write.table(
    rownames(L5_ET_rna_filtered), "../scenicplus/L5_ET/L5_ET_gene_names.tsv",
    col.names = FALSE, quote = FALSE, sep = "\t", row.names = FALSE)
