==> multiome/scenicplus/L5_ET/01.3_prep_atac_scanpy.ipynb <==
# Export ATAC to anndata

import scanpy as sc
import pandas as pd

import os

from pathlib import Path
root_dir = Path("/tscc/projects/ps-epigen/users/biy022/biccn/data/SNAREdata/scenicplus/")
os.chdir(root_dir / "L5_ET")

atac_counts = sc.read_mtx("L5_ET_atac_matrix.mtx")

atac_counts = atac_counts.T.copy()

cell_meta = pd.read_csv("L5_ET_meta_data.tsv", header=0, index_col=0, sep="\t")
region_names = pd.read_csv("L5_ET_atac_regions.tsv", header=None, index_col=0, sep="\t")
region_names.index.name = None

atac_counts.obs = cell_meta
atac_counts.var = region_names

atac_counts

atac_counts.write("L5_ET_atac_matrix.h5ad")

sum(atac_counts.X[0:100, 0:100].todense())
